+++ packages/styles/src/app/scss/ScssBorder.tsx
@@ -12,10 +12,6 @@ export const ScssBorder = () => {
         <div className="scss-border__radius scss-border__radius--4xl">Border Radius 4X-Large</div>
         <div className="scss-border__radius scss-border__radius--full">Border Radius Full</div>
       </div>
-      <div className="scss-border__width-wrapper">
-        <div className="scss-border__width scss-border__width--1">Border Width 1</div>
-        <div className="scss-border__width scss-border__width--2">Border Width 2</div>
-      </div>
     </div>
   );
 };